# UTF-8 support