--- a/src/Data/Text/Rope.hs
+++ b/src/Data/Text/Rope.hs
@@ -118,6 +118,7 @@ linesMetrics tl = Metrics
   }
 
 #ifdef DEBUG
+deriving instance Show Metrics
 deriving instance Show Rope
 #else
 instance Show Rope where