… bench/Main.hs
… bench/bench-utf8.txt
… changelog.md
… src/Data/Text/Utf16/Rope.hs
… src/Data/Text/Utf16/Rope/Mixed.hs
… src/Data/Text/Utf8/Lines.hs
… src/Data/Text/Utf8/Rope.hs
… test/Main.hs
… test/MixedRope.hs
… test/Utf8Lines.hs
… test/Utf8Rope.hs
… test/Utils.hs
… text-rope.cabal