Summary of changes:
* Don't store position metrics, but instead store newlines + char/code unit lengths. By not storing positions, we can subtract metrics and sometimes avoid some measurements.
* Add UTF-8 modules.
* Add UTF-8 operations to mixed rope.

Some benchmarks on my machine:

```
ghc-9.0.2 (UTF-16-native text)
  master
    Split at position
      Unicode
        text-rope:       OK
          7.11 ms ± 389 μs
        text-rope-mixed: OK
          7.49 ms ± 701 μs
      UTF-16
        text-rope:       OK
          2.71 ms ± 179 μs
        text-rope-mixed: OK
          6.46 ms ± 337 μs
    Split at offset
      Unicode
        text-rope:       OK
          12.6 ms ± 674 μs
        text-rope-mixed: OK
          13.8 ms ± 777 μs
      UTF-16
        text-rope:       OK
          2.63 ms ±  89 μs
        text-rope-mixed: OK
          6.66 ms ± 342 μs
  this branch
    Split at position
      Unicode
        text-rope:       OK
          4.05 ms ± 253 μs
        text-rope-mixed: OK
          3.89 ms ± 378 μs
      UTF-16
        text-rope:       OK
          3.50 ms ± 250 μs
        text-rope-mixed: OK
          3.81 ms ± 365 μs
      UTF-8
        text-rope:       OK
          3.17 ms ± 268 μs
        text-rope-mixed: OK
          3.68 ms ± 194 μs
    Split at offset
      Unicode
        text-rope:       OK
          2.36 ms ± 202 μs
        text-rope-mixed: OK
          2.52 ms ± 170 μs
      UTF-16
        text-rope:       OK
          2.11 ms ±  90 μs
        text-rope-mixed: OK
          2.45 ms ± 173 μs
      UTF-8
        text-rope:       OK
          1.97 ms ± 170 μs
        text-rope-mixed: OK
          2.40 ms ± 199 μs

ghc-9.4.8 (UTF-8-native text)
  master
    Split at position
      Unicode
        text-rope:       OK
          4.43 ms ± 353 μs
        text-rope-mixed: OK
          6.05 ms ± 588 μs
      UTF-16
        text-rope:       OK
          3.57 ms ± 190 μs
        text-rope-mixed: OK
          5.66 ms ± 221 μs
    Split at offset
      Unicode
        text-rope:       OK
          5.08 ms ± 361 μs
        text-rope-mixed: OK
          6.66 ms ± 359 μs
      UTF-16
        text-rope:       OK
          3.93 ms ± 200 μs
        text-rope-mixed: OK
          6.06 ms ± 397 μs
  this branch
    Split at position
      Unicode
        text-rope:       OK
          4.05 ms ± 253 μs
        text-rope-mixed: OK
          3.89 ms ± 378 μs
      UTF-16
        text-rope:       OK
          3.50 ms ± 250 μs
        text-rope-mixed: OK
          3.81 ms ± 365 μs
      UTF-8
        text-rope:       OK
          3.17 ms ± 268 μs
        text-rope-mixed: OK
          3.68 ms ± 194 μs
    Split at offset
      Unicode
        text-rope:       OK
          2.36 ms ± 202 μs
        text-rope-mixed: OK
          2.52 ms ± 170 μs
      UTF-16
        text-rope:       OK
          2.11 ms ±  90 μs
        text-rope-mixed: OK
          2.45 ms ± 173 μs
      UTF-8
        text-rope:       OK
          1.97 ms ± 170 μs
        text-rope-mixed: OK
          2.40 ms ± 199 μs
```

In summary, it's faster on almost every benchmark, but the pure UTF-16 rope is slightly slower with text-1 because it doesn't benefit from the measurement avoidance (since getting the UTF-16 length is then constant time).

Addresses https://github.com/Bodigrim/text-rope/issues/3.