@ollef 2024-03-26
Okay, I think I've addressed most comments.

A couple of notes and questions in line.