@Bodigrim 2024-04-28
@ollef do you need a release of `text-rope` soon to publish your changes?