@ollef 2024-04-29
@Bodigrim: Not urguently, but it might be nice to get it out there eventually. :) I can try using the Github version in the meantime.