@Bodigrim 2024-06-08
Thanks, I'll wait for #6 to land before making a release then.