@Bodigrim 2024-07-17
@ollef published as https://hackage.haskell.org/package/text-rope-0.3